Add LoRA-based LayerNorm removal with eval pipeline
- Full-rank LoRA adapters (A=identity, B=zeros) on all Conv1D layers
- Progressive FakeLayerNorm removal schedule reused from base codebase
- LR sweep configs for GPT-2 small (1e-4 to 3e-3)
- Scale-up configs for GPT-2 medium, large, XL at lr=6e-4
- Fix bf16 NaN issue: LoRA params now use float32
- Fix pile_eval bf16->numpy conversion
- Eval script (eval_lora.py): Pile-10k perplexity, HellaSwag, MMLU
- GPT-2 small results: Pile-10k PPL 16.3→25.9, HellaSwag -2pts

+++ pile_eval.py
@@ -139,7 +139,7 @@ def evaluate_model_on_pile(model, processed_examples, batch_size=8, device=None,
             if isinstance(model, HookedTransformer):
                 logits = model(batch_input_ids)
                 loss = model.loss_fn(logits, batch_input_ids, per_token=True)
-                loss_list.append(loss.flatten().cpu().numpy())
+                loss_list.append(loss.flatten().cpu().float().numpy())
                 batch_loss = loss.sum().item()
                 batch_tokens = loss.numel()
             else:
@@ -155,7 +155,7 @@ def evaluate_model_on_pile(model, processed_examples, batch_size=8, device=None,
                 loss_fct = torch.nn.CrossEntropyLoss(reduction='none')
                 per_token_loss = loss_fct(shift_logits.view(-1, shift_logits.size(-1)), shift_labels.view(-1))            
                 # Append to your loss list (assuming loss_list is defined elsewhere)
-                loss_list.append(per_token_loss.flatten().cpu().numpy())
+                loss_list.append(per_token_loss.flatten().cpu().float().numpy())
 
                 
                 # For HF models, get more accurate per-token loss
